Split extension-specific code out of src.digitaltwin
