+++ floodresilience/retrieve_static_boundaries.py
@@ -3,6 +3,7 @@
 This script automates the retrieval and storage of geospatial data from various providers using the 'geoapis' library.
 It populates the 'geospatial_layers' table in the database and stores user log information for tracking and reference.
 """
+import pathlib
 
 import geopandas as gpd
 
@@ -12,6 +13,7 @@ from src.digitaltwin.utils import LogLevel, setup_logging, get_catchment_area
 
 def main(
         selected_polygon_gdf: gpd.GeoDataFrame,
+        instruction_json_path: pathlib.Path,
         log_level: LogLevel = LogLevel.DEBUG) -> None:
     """
     Connect to various data providers to fetch geospatial data for the selected polygon, i.e., the catchment area.
@@ -22,6 +24,8 @@ def main(
     ----------
     selected_polygon_gdf : gpd.GeoDataFrame
         A GeoDataFrame representing the selected polygon, i.e., the catchment area.
+    instruction_json_path : pathlib.Path
+        The path to the instruction json file that specifies the geospatial data to be retrieved.
     log_level : LogLevel = LogLevel.DEBUG
         The log level to set for the root logger. Defaults to LogLevel.DEBUG.
         The available logging levels and their corresponding numeric values are:
@@ -40,8 +44,8 @@ def main(
     tables.define_custom_db_functions(engine)
     # Get catchment area.
     catchment_area = get_catchment_area(selected_polygon_gdf, to_crs=2193)
-    # Store 'static_boundary_instructions' records in the 'geospatial_layers' table in the database.
-    instructions_records_to_db.store_instructions_records_to_db(engine)
+    # Store records from instruction json in the 'geospatial_layers' table in the database.
+    instructions_records_to_db.store_instructions_records_to_db(engine, instruction_json_path)
     # Store geospatial layers data in the database
     data_to_db.store_geospatial_layers_data_to_db(engine, catchment_area)
     # Store user log information in the database
@@ -50,7 +54,9 @@ def main(
 
 if __name__ == "__main__":
     sample_polygon = gpd.GeoDataFrame.from_file("selected_polygon.geojson")
+    sample_instruction_path = pathlib.Path("floodresilience/static_boundary_instructions.json")
     main(
         selected_polygon_gdf=sample_polygon,
+        instruction_json_path=sample_instruction_path,
         log_level=LogLevel.DEBUG
     )
